put path-related utilities into Utilities.js
